backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
sqlalchemy>=2.0,<3.0
pydantic-settings>=2.3,<3.0
python-multipart>=0.0.9,<1.0
Pillow>=10.4,<15.0
faster-whisper>=1.0
huggingface_hub>=1.5,<2.0
httpx>=0.27,<1.0
